inserting fastas to database
beginning of a functioning fasta insert system works for both texts, files or both. it separates headers and sequences and inputs them into arrays where the index for fasta header x is the same for fasta sequences x etc.

Now i need to make sure the fasta and sequences are good, then fetch necessary information from header and then input into database correctly

+++ Entryhandling/performins.php
@@ -4,12 +4,10 @@ ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
 
 session_start();
-if(empty($_SESSION['loggedin'])){
-    header('Location:../index.php');
-}
-
-$iferror=0;
 $_SESSION['UploadError'] = '';
+
+
+//This section will handle file uploads
 
 // get details of the uploaded file
 $fileTmpPath = $_FILES['uploadedFile']['tmp_name'];
@@ -18,39 +16,57 @@ $fileSize = $_FILES['uploadedFile']['size'];
 $fileType = $_FILES['uploadedFile']['type'];
 $fileNameCmps = explode(".", $fileName);
 $fileExtension = strtolower(end($fileNameCmps));
-$_SESSION['type']=$_SESSION['type'] . $fileType;
 
-// sanitize file-name
-$newFileName = md5(time() . $fileName) . '.' . $fileExtension;
+if(!empty($fileSize)){
+    // Create a new temporary file name. 
+    $newFileName = md5(time() . $fileName) . '.' . $fileExtension;
 
-// check if file has one of the following extensions
-$allowedfileExtensions = array('txt', 'fasta');
+    //Select allowed formats. 
+    $allowedfileExtensions = array('txt', 'fasta');
 
-if (in_array($fileExtension, $allowedfileExtensions))
-{
-    // directory in which the uploaded file will be moved
-    $uploadFileDir = '.\uploadedfiles/';
-    $dest_path = $uploadFileDir . $newFileName;
-
-    if(move_uploaded_file($fileTmpPath, $dest_path)) 
+    //check to see if uploaded file has correct input format. 
+    if (in_array($fileExtension, $allowedfileExtensions))
     {
-        $ffile = file_get_contents($dest_path);
-        echo $ffile;
-        unlink($dest_path);
-    }else
-    {
-        $_SESSION['UploadError'] = $_SESSION['UploadError'] . '<br>' . 'File could not be moved to correct destination';
+        //select where to put the file
+        $uploadFileDir = '.\uploadedfiles/';
+        $dest_path = $uploadFileDir . $newFileName;
+
+    //if upload was succesful, fetch it as a string and then remove it. 
+        if(move_uploaded_file($fileTmpPath, $dest_path)) 
+        {
+            $ffile = file_get_contents($dest_path);
+            unlink($dest_path);
+        }else
+        {
+            $_SESSION['UploadError'] = $_SESSION['UploadError'] . '<br>' . 'File could not be moved to correct destination';
+        }
+    }else {
+        $_SESSION['UploadError'] = $_SESSION['UploadError'] . '<br>' . 'Only .fasta fileformats allowed';
     }
-}else {
-    $_SESSION['UploadError'] = $_SESSION['UploadError'] . '<br>' . 'Only .fasta fileformats allowed';
+
+    include 'filefastahandler.php';
 }
 
+//This section will handle text inputs
+if(!empty($_POST['fastatext'])){
+    $ftext = $_POST['fastatext'];
+    include 'textfastahandler.php';
+}
+elseif(empty($ftext) && empty($fileSize)){
+    $_SESSION['UploadError'] = $_SESSION['UploadError'] . '<br>' . 'Empty submission';
+}
 
-
+//If there was an error, bring it back to the upload page and display the error. 
 if (!empty($_SESSION['UploadError'])){
     $_SESSION['UploadError'] ='File Upload Error: ' . $_SESSION['UploadError'];
-    header('Location:./insertDB.php');
+    header('Location:./insertform.php');
 }
+
+include 'fastadivider.php';
+print_r($header);
+echo '<br><br>';
+print_r($seq);
+include 'headerhandler.php';
 
 ?>
